+++ weapon.h
@@ -5,8 +5,8 @@
 
 using namespace std;
 
-void text(string szText,int X,int Y,WORD color);
-void num(int num,int X,int Y,WORD color);
+void text(string szText,short X,short Y,WORD color);
+void num(int num, short X, short Y,WORD color);
 
 class Weapon : public Item{
 public:
